+++ MolecularNodes/__init__.py
@@ -17,7 +17,7 @@ bl_info = {
     "author"      : "Brady Johnston", 
     "description" : "Importer and nodes for working with structural biology data in Blender.",
     "blender"     : (3, 4, 0),
-    "version"     : (2, 3, 1),
+    "version"     : (2, 4, 1),
     "location"    : "Scene Properties -> MolecularNodes",
     "warning"     : "",
     "doc_url"     : "https://bradyajohnston.github.io/MolecularNodes/", 
@@ -27,15 +27,21 @@ bl_info = {
 
 import bpy
 from . import pkg
-from .pref import *
 pkg.verify()
-from .load import *
 from .ui import *
 from .md import *
-
+from .pkg import *
 
 
 def register():
+    bpy.types.Scene.pypi_mirror = bpy.props.StringProperty(
+        name = 'pypi_mirror', 
+        description = 'PyPI Mirror', 
+        options = {'TEXTEDIT_UPDATE'}, 
+        default = 'Default', 
+        subtype = 'NONE', 
+        search = get_pypi_mirror_alias,
+        )
     bpy.types.Scene.mol_pdb_code = bpy.props.StringProperty(
         name = 'pdb_code', 
         description = 'The 4-character PDB code to download', 
@@ -138,7 +144,7 @@ def register():
     )
 
     bpy.utils.register_class(TrajectorySelectionList)
-    bpy.utils.register_class(TrajectorySelectionListUI)
+    bpy.utils.register_class(MOL_UL_TrajectorySelectionListUI)
     bpy.utils.register_class(TrajectorySelection_OT_NewItem)
     bpy.utils.register_class(TrajectorySelection_OT_DeleteIem)
     
@@ -151,9 +157,8 @@ def register():
     )
     
     bpy.types.NODE_MT_add.append(mol_add_node_menu)
-    
+
     bpy.utils.register_class(MOL_PT_panel)
-    bpy.utils.register_class(MOL_PT_AddonPreferences)
     bpy.utils.register_class(MOL_MT_Add_Node_Menu)
     bpy.utils.register_class(MOL_MT_Add_Node_Menu_Properties)
     bpy.utils.register_class(MOL_MT_Add_Node_Menu_Styling)
@@ -168,7 +173,9 @@ def register():
     bpy.utils.register_class(MOL_MT_Default_Style)
 
     bpy.utils.register_class(MOL_OT_Style_Surface_Custom)
+
     bpy.utils.register_class(MOL_OT_Import_Protein_RCSB)
+
     bpy.utils.register_class(MOL_OT_Import_Method_Selection)
     bpy.utils.register_class(MOL_OT_Import_Protein_Local)
     bpy.utils.register_class(MOL_OT_Import_Protein_MD)
@@ -177,10 +184,11 @@ def register():
     bpy.utils.register_class(MOL_OT_Color_Chain)
     bpy.utils.register_class(MOL_OT_Chain_Selection_Custom)
     bpy.utils.register_class(MOL_OT_Ligand_Selection_Custom)
-    
     bpy.utils.register_class(MOL_OT_install_dependencies)
+
     bpy.utils.register_class(MOL_OT_Add_Custom_Node_Group)
 
+    bpy.utils.register_class(MOL_OT_Residues_Selection_Custom)
 
 
 def unregister():
@@ -206,12 +214,11 @@ def unregister():
     bpy.types.NODE_MT_add.remove(mol_add_node_menu)
     
     bpy.utils.unregister_class(TrajectorySelectionList)
-    bpy.utils.unregister_class(TrajectorySelectionListUI)
+    bpy.utils.unregister_class(MOL_UL_TrajectorySelectionListUI)
     bpy.utils.unregister_class(TrajectorySelection_OT_NewItem)
     bpy.utils.unregister_class(TrajectorySelection_OT_DeleteIem)
 
     bpy.utils.unregister_class(MOL_PT_panel)
-    bpy.utils.unregister_class(MOL_PT_AddonPreferences)
     bpy.utils.unregister_class(MOL_MT_Add_Node_Menu)
     bpy.utils.unregister_class(MOL_MT_Add_Node_Menu_Properties)
     bpy.utils.unregister_class(MOL_MT_Add_Node_Menu_Styling)
@@ -236,9 +243,10 @@ def unregister():
     bpy.utils.unregister_class(MOL_OT_Color_Chain)
     bpy.utils.unregister_class(MOL_OT_Chain_Selection_Custom)
     
-    bpy.utils.unregister_class(MOL_OT_install_dependencies)
     bpy.utils.unregister_class(MOL_OT_Add_Custom_Node_Group)
+    bpy.utils.unregister_class(MOL_OT_install_dependencies)
 
+    bpy.utils.unregister_class(MOL_OT_Residues_Selection_Custom)
 
 if __name__=="__main__":
     register()