+++ MolecularNodes/ui.py
@@ -1,12 +1,9 @@
 import bpy
-from .pref import *
-from .tools import property_exists
 from . import nodes
 from . import pkg
 from . import load
 from . import md
 from . import assembly
-import os
 
 
 
@@ -39,8 +36,6 @@ class MOL_OT_Import_Protein_RCSB(bpy.types.Operator):
 
     def invoke(self, context, event):
         return self.execute(context)
-
-
 
 
 # operator that calls the function to import the structure from a local file
@@ -115,7 +110,11 @@ class MOL_OT_Import_Protein_MD(bpy.types.Operator):
             starting_style = bpy.context.scene.mol_import_default_style
             )
         bpy.context.view_layer.objects.active = mol_object
-        self.report({'INFO'}, message=f"Imported '{file_top}' as {mol_object.name} with {str(n_frames)} frames from '{file_traj}'.")
+        self.report(
+            {'INFO'}, 
+            message=f"Imported '{file_top}' as {mol_object.name} with {str(n_frames)} \
+                frames from '{file_traj}'."
+                )
         
         return {"FINISHED"}
 
@@ -142,8 +141,10 @@ def MOL_PT_panel_local(layout_function, ):
     col_main.active = True
     col_main.label(text = "Open Local File")
     row_name = col_main.row(align = False)
-    row_name.prop(bpy.context.scene, 'mol_import_local_name', text = "Name", icon_value = 0, emboss = True)
-    row_name.operator('mol.import_protein_local', text = "Load", icon='FILE_TICK', emboss = True)
+    row_name.prop(bpy.context.scene, 'mol_import_local_name', 
+                    text = "Name", icon_value = 0, emboss = True)
+    row_name.operator('mol.import_protein_local', text = "Load", 
+                        icon='FILE_TICK', emboss = True)
     row_import = col_main.row()
     row_import.prop(
         bpy.context.scene, 'mol_import_local_path', 
@@ -204,8 +205,8 @@ def MOL_PT_panel_md_traj(layout_function, scene):
     row = col_main.row(align=True)
     
     row = row.split(factor = 0.9)
-    row.template_list('TrajectorySelectionListUI', 'A list', scene, 
-                      "trajectory_selection_list", scene, "list_index", rows=3)
+    row.template_list('MOL_UL_TrajectorySelectionListUI', 'A list', scene, 
+                        "trajectory_selection_list", scene, "list_index", rows=3)
     col = row.column()
     col.operator('trajectory_selection_list.new_item', icon="ADD", text="")
     col.operator('trajectory_selection_list.delete_item', icon="REMOVE", text="")
@@ -224,7 +225,12 @@ class MOL_OT_Import_Method_Selection(bpy.types.Operator):
     bl_label = "import_method"
     bl_description = "Change Structure Import Method"
     bl_options = {"REGISTER", "UNDO"}
-    mol_interface_value: bpy.props.IntProperty(name = 'interface_value', description = '', default = 0, subtype = 'NONE')
+    mol_interface_value: bpy.props.IntProperty(
+        name = 'interface_value', 
+        description = '', 
+        default = 0, 
+        subtype = 'NONE'
+        )
 
     @classmethod
     def poll(cls, context):
@@ -289,17 +295,37 @@ class MOL_MT_Default_Style(bpy.types.Menu):
 def MOL_PT_panel_ui(layout_function, scene): 
     layout_function.label(text = "Import Options", icon = "MODIFIER")
     if not pkg.available():
+        
+        col_main = layout_function.column(heading = '', align = False)
+        col_main.alert = False
+        col_main.enabled = True
+        col_main.active = True
+        col_main.use_property_split = False
+        col_main.use_property_decorate = False
+        col_main.scale_x = 1.0
+        col_main.scale_y = 1.0
+        
+        col_main.alignment = 'Expand'.upper()
+        col_main.label(text = "Set PyPI Mirror")
+        row_import = col_main.row()
+        row_import.prop(bpy.context.scene, 'pypi_mirror',text='PyPI')
         layout_function.operator('mol.install_dependencies', text = 'Install Packages')
+        
     else:
         box = layout_function.box()
         grid = box.grid_flow(columns = 2)
         
-        grid.prop(bpy.context.scene, 'mol_import_center', text = 'Centre Structre', icon_value=0, emboss=True)
-        grid.prop(bpy.context.scene, 'mol_import_del_solvent', text = 'Delete Solvent', icon_value=0, emboss=True)
-        grid.prop(bpy.context.scene, 'mol_import_include_bonds', text = 'Import Bonds', icon_value=0, emboss=True)
+        grid.prop(bpy.context.scene, 'mol_import_center', 
+                    text = 'Centre Structre', icon_value=0, emboss=True)
+        grid.prop(bpy.context.scene, 'mol_import_del_solvent', 
+                    text = 'Delete Solvent', icon_value=0, emboss=True)
+        grid.prop(bpy.context.scene, 'mol_import_include_bonds', 
+                    text = 'Import Bonds', icon_value=0, emboss=True)
         grid.menu(
             'MOL_MT_Default_Style', 
-            text = ['Atoms', 'Ribbon', 'Ball and Stick'][bpy.context.scene.mol_import_default_style])
+            text = ['Atoms', 'Ribbon', 'Ball and Stick'][
+                bpy.context.scene.mol_import_default_style
+                ])
         box = layout_function
         row = box.row(heading = '', align=True)
         row.alignment = 'EXPAND'
@@ -316,6 +342,8 @@ def MOL_PT_panel_ui(layout_function, scene):
             MOL_PT_panel_local(layout_function)
         else:
             MOL_PT_panel_md_traj(layout_function, scene)
+
+
 
 class MOL_PT_panel(bpy.types.Panel):
     bl_label = 'Molecular Nodes'
@@ -343,17 +371,21 @@ def mol_add_node(node_name):
     prev_context = bpy.context.area.type
     bpy.context.area.type = 'NODE_EDITOR'
     # actually invoke the operator to add a node to the current node tree
-    # use_transform=True ensures it appears where the user's mouse is and is currently being moved
-    # so the user can place it where they wish
-    bpy.ops.node.add_node('INVOKE_DEFAULT', type='GeometryNodeGroup', use_transform=True)
+    # use_transform=True ensures it appears where the user's mouse is and is currently 
+    # being moved so the user can place it where they wish
+    bpy.ops.node.add_node(
+        'INVOKE_DEFAULT', 
+        type='GeometryNodeGroup', 
+        use_transform=True
+        )
     bpy.context.area.type = prev_context
     bpy.context.active_node.node_tree = bpy.data.node_groups[node_name]
     bpy.context.active_node.width = 200.0
-    # checks to see if the node as a 'Material' property, and if it does, set MOL_atomic_materic as that property
-    if (property_exists("bpy.data.node_groups[bpy.context.active_object.modifiers.active.node_group.name].nodes[bpy.context.active_node.name].inputs['Material'].default_value", globals(), locals())):
-        mat = nodes.mol_base_material()
-        bpy.data.node_groups[bpy.context.active_object.modifiers.active.node_group.name].nodes[bpy.context.active_node.name].inputs['Material'].default_value = bpy.data.materials[mat.name]
-    
+    # if added node has a 'Material' input, set it to the default MN material
+    input_mat = bpy.context.active_node.inputs.get('Material')
+    if input_mat:
+        input_mat = nodes.mol_base_material().name
+
 class MOL_OT_Add_Custom_Node_Group(bpy.types.Operator):
     bl_idname = "mol.add_custom_node_group"
     bl_label = "Add Custom Node Group"
@@ -386,7 +418,8 @@ class MOL_OT_Add_Custom_Node_Group(bpy.types.Operator):
             nodes.mol_append_node(self.node_name)
             mol_add_node(self.node_name)
         except RuntimeError:
-            self.report({'ERROR'}, message='Failed to add node. Ensure you are not in edit mode.')
+            self.report({'ERROR'}, 
+                        message='Failed to add node. Ensure you are not in edit mode.')
         return {"FINISHED"}
     
     def invoke(self, context, event):
@@ -397,7 +430,8 @@ def menu_item_interface(layout_function,
                         label, 
                         node_name, 
                         node_description='Add custom MolecularNodes node group.'):
-    op = layout_function.operator('mol.add_custom_node_group', text = label, emboss = True, depress=False)
+    op=layout_function.operator('mol.add_custom_node_group', 
+                                text = label, emboss = True, depress=False)
     op.node_name = node_name
     op.node_description = node_description
 
@@ -405,7 +439,9 @@ def menu_item_interface(layout_function,
 class MOL_OT_Style_Surface_Custom(bpy.types.Operator):
     bl_idname = "mol.style_surface_custom"
     bl_label = "My Class Name"
-    bl_description = "Create a split surface representation.\nGenerates an isosurface based on atomic vdw_radii. Each chain has its own separate surface representation"
+    bl_description = "Create a split surface representation.\nGenerates an isosurface \
+        based on atomic vdw_radii. Each chain has its own separate surface \
+        representation"
     bl_options = {"REGISTER", "UNDO"}
     
     @classmethod
@@ -429,7 +465,10 @@ class MOL_OT_Style_Surface_Custom(bpy.types.Operator):
 class MOL_OT_Assembly_Bio(bpy.types.Operator):
     bl_idname = "mol.assembly_bio"
     bl_label = "Build"
-    bl_description = "**PDB Downloaded Structures Only**\nAdds node to build biological assembly based on symmetry operations that are extraced from the structure file. Currently this is only supported for structures that were downloaded from the PDB"
+    bl_description = "**PDB Downloaded Structures Only**\nAdds node to build \
+        biological assembly based on symmetry operations that are extraced from the \
+        structure file. Currently this is only supported for structures that were \
+        downloaded from the PDB"
     bl_options = {"REGISTER", "UNDO"}
     
     @classmethod
@@ -445,13 +484,25 @@ class MOL_OT_Assembly_Bio(bpy.types.Operator):
             )
         except:
             node_bio_assembly = None
-            self.report({'WARNING'}, message = 'Unable to detect biological assembly information.')
+            self.report(
+                {'WARNING'}, 
+                message = 'Unable to detect biological assembly information.'
+                )
         
         if node_bio_assembly:
             mol_add_node(node_bio_assembly.name)
         
         return {"FINISHED"}
 
+def menu_residues_selection_custom(layout_function):
+    obj = bpy.context.view_layer.objects.active
+    label = 'Res ID'
+    op = layout_function.operator(
+        'mol.residues_selection_custom', 
+        text = label, 
+        emboss = True, 
+        depress = True
+    )
 
 def menu_item_surface_custom(layout_function, label):
     op = layout_function.operator('mol.style_surface_custom', 
@@ -468,7 +519,8 @@ def menu_item_color_chains(layout_function, label):
 class MOL_OT_Color_Chain(bpy.types.Operator):
     bl_idname = "mol.color_chains"
     bl_label = "My Class Name"
-    bl_description = "Create a custom node for coloring each chain of a structure individually.\nRequires chain information to be available from the structure"
+    bl_description = "Create a custom node for coloring each chain of a structure \
+        individually.\nRequires chain information to be available from the structure"
     bl_options = {"REGISTER", "UNDO"}
     
     @classmethod
@@ -501,7 +553,9 @@ def menu_chain_selection_custom(layout_function):
 class MOL_OT_Chain_Selection_Custom(bpy.types.Operator):
     bl_idname = "mol.chain_selection_custom"
     bl_label = "Chain Selection"
-    bl_description = "Create a selection based on the chains.\nThis node is built on a per-molecule basis, taking into account the chain_ids that were detected. If no chain information is available this node will not work"
+    bl_description = "Create a selection based on the chains.\nThis node is built on a \
+        per-molecule basis, taking into account the chain_ids that were detected. If \
+        no chain information is available this node will not work"
     bl_options = {"REGISTER", "UNDO"}
     
     @classmethod
@@ -522,6 +576,41 @@ class MOL_OT_Chain_Selection_Custom(bpy.types.Operator):
         
         return {"FINISHED"}
 
+
+class MOL_OT_Residues_Selection_Custom(bpy.types.Operator):
+    bl_idname = "mol.residues_selection_custom"
+    bl_label = "Multiple Residue Selection"
+    bl_description = "Create a selection based on the provided residue strings.\nThis \
+        node is built on a per-molecule basis, taking into account the residues that \
+        were input."
+    bl_options = {"REGISTER", "UNDO"}
+
+    input_resid_string: bpy.props.StringProperty(
+        name="Select residue IDs: ",
+        description="Enter a string value.",
+        default="19,94,1-16"
+    )
+
+    @classmethod
+    def poll(cls, context):
+        return True
+    
+    def execute(self, context):
+        obj = bpy.context.view_layer.objects.active
+        node_residues = nodes.resid_multiple_selection(
+            node_name = 'MOL_sel_residues', 
+            input_resid_string = self.input_resid_string, 
+            )
+    
+        
+        mol_add_node(node_residues.name)
+        return {"FINISHED"}
+
+    def invoke(self, context, event):
+        return context.window_manager.invoke_props_dialog(self)
+
+
+
 def menu_ligand_selection_custom(layout_function):
     obj = bpy.context.view_layer.objects.active
     label = 'Ligands ' + str(obj.name)
@@ -535,7 +624,9 @@ def menu_ligand_selection_custom(layout_function):
 class MOL_OT_Ligand_Selection_Custom(bpy.types.Operator):
     bl_idname = "mol.ligand_selection_custom"
     bl_label = "Ligand Selection"
-    bl_description = "Create a selection based on the ligands.\nThis node is built on a per-molecule basis, taking into account the chain_ids that were detected. If no chain information is available this node will not work"
+    bl_description = "Create a selection based on the ligands.\nThis node is built on \
+        a per-molecule basis, taking into account the chain_ids that were detected. If \
+        no chain information is available this node will not work"
     bl_options = {"REGISTER", "UNDO"}
     
     @classmethod
@@ -585,17 +676,19 @@ class MOL_MT_Add_Node_Menu_Color(bpy.types.Menu):
                             "Sets a new color for the selected atoms")
         layout.separator()
         menu_item_interface(layout, 'Goodsell Colors', 'MOL_color_goodsell', 
-                            "Adjusts the given colors to copy the 'Goodsell Style'.\n" + 
-                            "Darkens the non-carbon atoms and keeps the carbon atoms the same color. " +
-                            "Highlights differences without being too visually busy")
+                            "Adjusts the given colors to copy the 'Goodsell Style'.\n \
+                            Darkens the non-carbon atoms and keeps the carbon atoms \
+                            the same color. Highlights differences without being too \
+                            visually busy")
         layout.separator()
-        menu_item_interface(layout, 'Color by Atomic Number', 'MOL_color_atomic_number', 
+        menu_item_interface(layout, 'Color by Atomic Number', 'MOL_color_atomic_number',
                             "Creates a color based on atomic_number field")
         menu_item_interface(layout, 'Color by Element', 'MOL_color_element', 
                             "Choose a color for each of the first 20 elements")
         menu_item_color_chains(layout, 'Color by Chains')
         menu_item_interface(layout, 'Color Atomic', 'MOL_style_color', 
-                            "Choose a color for the most common elements in PDB structures")
+                            "Choose a color for the most common elements in PDB \
+                            structures")
 
 class MOL_MT_Add_Node_Menu_Bonds(bpy.types.Menu):
     bl_idname = 'MOL_MT_ADD_NODE_MENU_BONDS'
@@ -609,13 +702,16 @@ class MOL_MT_Add_Node_Menu_Bonds(bpy.types.Menu):
         layout = self.layout
         layout.operator_context = "INVOKE_DEFAULT"
         menu_item_interface(layout, 'Find Bonds', 'MOL_bonds_find', 
-                            "Finds bonds between atoms based on distance.\n" + 
-                            "Based on the vdw_radii for each point, finds other points within a certain radius to create a bond to. " + 
-                            "Does not preserve the index for the points. Does not detect bond type")
+                            "Finds bonds between atoms based on distance.\n\
+                            Based on the vdw_radii for each point, finds other points \
+                            within a certain radius to create a bond to. Does not \
+                            preserve the index for the points. Does not detect bond type")
         menu_item_interface(layout, 'Break Bonds', 'MOL_bonds_break', 
-                            "Will delete a bond between atoms that already exists based on a distance cutoff")
+                            "Will delete a bond between atoms that already exists \
+                            based on a distance cutoff")
         menu_item_interface(layout, 'Find Bonded Atoms', 'MOL_bonds_find_bonded', 
-                            "Based on an initial selection, finds atoms which are within a certain number of bonds away")
+                            "Based on an initial selection, finds atoms which are \
+                            within a certain number of bonds away")
 
 class MOL_MT_Add_Node_Menu_Styling(bpy.types.Menu):
     bl_idname = 'MOL_MT_ADD_NODE_MENU_SYLING'
@@ -629,21 +725,29 @@ class MOL_MT_Add_Node_Menu_Styling(bpy.types.Menu):
         layout = self.layout
         layout.operator_context = "INVOKE_DEFAULT"
         menu_item_interface(layout, 'Atoms Cycles', 'MOL_style_atoms_cycles', 
-                            'A sphere atom representation, visible ONLY in Cycles. Based on point-cloud rendering')
+                            'A sphere atom representation, visible ONLY in Cycles. \
+                            Based on point-cloud rendering')
         menu_item_interface(layout, 'Atoms EEVEE', 'MOL_style_atoms_eevee', 
-                            'A sphere atom representation, visible in EEVEE and Cycles. Based on mesh instancing which slows down viewport performance')
+                            'A sphere atom representation, visible in EEVEE and \
+                            Cycles. Based on mesh instancing which slows down viewport \
+                            performance')
         menu_item_interface(layout, 'Ribbon Protein', 'MOL_style_ribbon_protein', 
-                            'Create a ribbon mesh based off of the alpha-carbons of the structure')
+                            'Create a ribbon mesh based off of the alpha-carbons of \
+                            the structure')
         menu_item_interface(layout, 'Ribbon Nucleic', 'MOL_style_ribbon_nucleic', 
-                            'Create a ribbon mesh and instanced cylinders for nucleic acids.')
+                            'Create a ribbon mesh and instanced cylinders for nucleic \
+                            acids.')
         menu_item_interface(layout, 'Surface', 'MOL_style_surface_single', 
-                            'Create a single joined surface representation.\n' +
-                            'Generates an isosurface based on atomic vdw_radii. All chains are part of the same surface. Use "Surface Split Chains" ' + 
-                            'to have a single surface per chain')
+                            "Create a single joined surface representation. \
+                            Generates an isosurface based on atomic vdw_radii. All \
+                            chains are part of the same surface. Use Surface Split \
+                            Chains to have a single surface per chain")
         menu_item_surface_custom(layout, 'Surface Split Chains')
         menu_item_interface(layout, 'Ball and Stick', 'MOL_style_ball_and_stick', 
-                            'A style node to create ball and stick representation.\n' +
-                            'Icospheres are instanced on atoms and cylinders for bonds. Bonds can be detected if they are not present in the structure')
+                            "A style node to create ball and stick representation. \
+                            Icospheres are instanced on atoms and cylinders for bonds. \
+                            Bonds can be detected if they are not present in the \
+                            structure")
 
 
 class MOL_MT_Add_Node_Menu_Selections(bpy.types.Menu):
@@ -659,7 +763,8 @@ class MOL_MT_Add_Node_Menu_Selections(bpy.types.Menu):
         layout.operator_context = "INVOKE_DEFAULT"
         menu_item_interface(layout, 'Select Atoms', 'MOL_sel_atoms', 
                             "Separate atoms based on a selection field.\n" +
-                            "Takes atoms and splits them into the selected atoms the inverted atoms, based on a selection field")
+                            "Takes atoms and splits them into the selected atoms the \
+                            inverted atoms, based on a selection field")
         menu_item_interface(layout, 'Separate Polymers', 'MOL_sel_sep_polymers', 
                             "Separate the Geometry into the different polymers.\n" + 
                             "Outputs for protein, nucleic & sugars")
@@ -668,32 +773,41 @@ class MOL_MT_Add_Node_Menu_Selections(bpy.types.Menu):
         menu_ligand_selection_custom(layout)
         layout.separator()
         menu_item_interface(layout, 'Atom Properties', 'MOL_sel_atom_propeties', 
-                            "Create a selection based on the properties of the atom.\n" + 
-                            "Fields for is_alpha_carbon, is_backbone, is_peptide, is_nucleic, is_solvent and is_carb")
+                            "Create a selection based on the properties of the atom.\n\
+                            Fields for is_alpha_carbon, is_backbone, is_peptide, \
+                            is_nucleic, is_solvent and is_carb")
         menu_item_interface(layout, 'Atomic Number', 'MOL_sel_atomic_number', 
-                            "Create a selection if input value equal to the atomic_number field.")
+                            "Create a selection if input value equal to the \
+                            atomic_number field.")
         menu_item_interface(layout, 'Element Name', 'MOL_sel_element_name', 
-                            "Create a selection of particular elements by name. Only first 20 elements supported")
+                            "Create a selection of particular elements by name. Only \
+                            first 20 elements supported")
         layout.separator()
         menu_item_interface(layout, 'Distance', 'MOL_sel_distance', 
-                            "Create a selection based on the distance to a selected object.\n" + 
-                            "The cutoff is scaled based on the objects scale and the 'Scale Cutoff' value.")
+                            "Create a selection based on the distance to a selected \
+                            object.\n The cutoff is scaled based on the objects scale \
+                            and the 'Scale Cutoff' value.")
         menu_item_interface(layout, 'Slice', 'MOL_sel_slice', 
-                            "Create a selection that is a slice along one of the XYZ axes, based on the position of an object.")
+                            "Create a selection that is a slice along one of the XYZ \
+                            axes, based on the position of an object.")
         layout.separator()
-        menu_item_interface(layout, 'Res Atoms', 'MOL_sel_res_atoms', 
-                            "Create a selection based on the atoms of a residue.\n" +
-                            "Selections for CA, backbone atoms (N, C, O), sidechain and backbone")
-        menu_item_interface(layout, 'Res Name', 'MOL_sel_res_name', 
+        menu_residues_selection_custom(layout)                        
+        menu_item_interface(layout, 'Res ID Single', 'MOL_sel_res_id', 
+                            "Create a selection if res_id matches input field")
+        menu_item_interface(layout, 'Res ID Range', 'MOL_sel_res_id_range', 
+                            "Create a selection if the res_id is within the given \
+                            thresholds")
+        menu_item_interface(layout, 'Res Name Peptide', 'MOL_sel_res_name', 
                             "Create a selection of particular amino acids by name")
         menu_item_interface(layout, 'Res Name Nucleic', 'MOL_sel_res_name_nucleic', 
                             "Create a selection of particular nucleic acids by name")
-        menu_item_interface(layout, 'Res ID', 'MOL_sel_res_id', 
-                            "Create a selection if res_id matches input field")
-        menu_item_interface(layout, 'Res ID Range', 'MOL_sel_res_id_range', 
-                            "Create a selection if the res_id is within the given thresholds")
         menu_item_interface(layout, 'Res Whole', 'MOL_sel_res_whole', 
-                            "Expand the selection to every atom in a residue, if any of those atoms are in the initial selection")
+                            "Expand the selection to every atom in a residue, if any \
+                            of those atoms are in the initial selection")
+        menu_item_interface(layout, 'Res Atoms', 'MOL_sel_res_atoms', 
+                            "Create a selection based on the atoms of a residue.\n" +
+                            "Selections for CA, backbone atoms (N, C, O), sidechain \
+                            and backbone")
 
 class MOL_MT_Add_Node_Menu_Assembly(bpy.types.Menu):
     bl_idname = 'MOL_MT_ADD_NODE_MENU_ASSEMBLY'
@@ -706,9 +820,14 @@ class MOL_MT_Add_Node_Menu_Assembly(bpy.types.Menu):
     def draw(self, context):
         layout = self.layout
         layout.operator_context = "INVOKE_DEFAULT"
-        layout.operator("mol.assembly_bio", text = "Biological Assembly", emboss = True, depress=True)
+        layout.operator("mol.assembly_bio", 
+                        text = "Biological Assembly", 
+                        emboss = True, 
+                        depress=True
+                        )
         menu_item_interface(layout, 'Center Assembly', 'MOL_assembly_center', 
-                            "Center the structure on the world origin based on bounding box")
+                            "Center the structure on the world origin based on \
+                            bounding box")
 
 class MOL_MT_Add_Node_Menu_Membranes(bpy.types.Menu):
     bl_idname = 'MOL_MT_ADD_NODE_MENU_MEMBRANES'
@@ -736,17 +855,21 @@ class MOL_MT_Add_Node_Menu_DNA(bpy.types.Menu):
         layout.operator_context = "INVOKE_DEFAULT"
         menu_item_interface(layout, 'Double Helix', 'MOL_dna_double_helix', 
                             "Create a DNA double helix from an input curve.\n" + 
-                            "Takes an input curve and instances for the bases, returns instances of the bases in a double helix formation")
+                            "Takes an input curve and instances for the bases, returns \
+                            instances of the bases in a double helix formation")
         menu_item_interface(layout, 'Bases', 'MOL_dna_bases', 
-                            "Provide the DNA bases as instances to be styled and passed onto the Double Helix node")
+                            "Provide the DNA bases as instances to be styled and \
+                            passed onto the Double Helix node")
         layout.separator()
         menu_item_interface(layout, 'Style Atoms Cyeles', 'MOL_dna_style_atoms_cycles', 
                             "Style the DNA bases with spheres only visible in Cycles")
         menu_item_interface(layout, 'Style Atoms EEVEE', 'MOL_dna_style_atoms_eevee', 
-                            "Style the DNA bases with spheres visible in Cycles and EEVEE")
+                            "Style the DNA bases with spheres visible in Cycles and \
+                            EEVEE")
         menu_item_interface(layout, 'Style Surface', 'MOL_dna_style_surface', 
                             "Style the DNA bases with surface representation")
-        menu_item_interface(layout, 'Style Ball and Stick', 'MOL_dna_style_ball_and_stick', 
+        menu_item_interface(layout, 'Style Ball and Stick', 
+                            'MOL_dna_style_ball_and_stick', 
                             "Style the DNA bases with ball and stick representation")
 
 class MOL_MT_Add_Node_Menu_Animation(bpy.types.Menu):
@@ -762,22 +885,33 @@ class MOL_MT_Add_Node_Menu_Animation(bpy.types.Menu):
         layout.operator_context = "INVOKE_DEFAULT"
         menu_item_interface(layout, 'Animate Frames', 'MOL_animate_frames', 
                             "Interpolate between frames of a trajectory." + 
-                            "Given a collection of frames for a trajectory, this node interpolates between them from start to finish based on the Animate field taking a value from 0 to 1. The positions of the Atoms are then moved based on this field")
+                            "Given a collection of frames for a trajectory, this node \
+                            interpolates between them from start to finish based on \
+                            the Animate field taking a value from 0 to 1. The \
+                            positions of the Atoms are then moved based on this field")
         menu_item_interface(layout, 'Animate Field', 'MOL_animate_field')
         menu_item_interface(layout, 'Animate Value', 'MOL_animate_value', 
-                            "Animates between given start and end values, based on the input start and end frame of the timeline. Clamped will limit the output to the 'To Min' and 'To Max', while unclamped will continue to interpolate past these values. 'Smoother Step' will ease in and out of these values, with default being linear interpolation")
+                            "Animates between given start and end values, based on \
+                            the input start and end frame of the timeline. Clamped \
+                            will limit the output to the 'To Min' and 'To Max', while \
+                            unclamped will continue to interpolate past these values. \
+                            'Smoother Step' will ease in and out of these values, with \
+                            default being linear interpolation")
         layout.separator()
         menu_item_interface(layout, 'Res Wiggle', "MOL_animate_res_wiggle", 
-                            "Wiggles the side chains of amino acids based on b_factor, adding movement to a structure.")
+                            "Wiggles the side chains of amino acids based on b_factor, \
+                            adding movement to a structure.")
         menu_item_interface(layout, 'Res to Curve', "MOL_animate_res_to_curve", 
-                            "Takes atoms and maps them along a curve, as a single long peptide chain.")
+                            "Takes atoms and maps them along a curve, as a single \
+                            long peptide chain.")
         layout.separator()
         menu_item_interface(layout, 'Noise Position', 'MOL_noise_position', 
                             "Generate 3D noise field based on the position attribute")
         menu_item_interface(layout, 'Noise Field', 'MOL_noise_field', 
                             "Generate a 3D noise field based on the given field")
         menu_item_interface(layout, 'Noise Repeat', 'MOL_noise_repeat', 
-                            "Generate a 3D noise field that repeats, based on the given field")
+                            "Generate a 3D noise field that repeats, based on the \
+                            given field")
 
 class MOL_MT_Add_Node_Menu_Utilities(bpy.types.Menu):
     bl_idname = 'MOL_MT_ADD_NODE_MENU_UTILITIES'
@@ -805,16 +939,22 @@ class MOL_MT_Add_Node_Menu(bpy.types.Menu):
     def draw(self, context):
         layout = self.layout.column_flow(columns=1)
         layout.operator_context = "INVOKE_DEFAULT"
-        # layout.menu('MOL_MT_ADD_NODE_MENU_PROPERTIES', text='Properties', icon_value=201)
-        layout.menu('MOL_MT_ADD_NODE_MENU_SYLING', text='Style', icon_value=77)
-        layout.menu('MOL_MT_ADD_NODE_MENU_COLOR', text='Color', icon = 'COLORSET_07_VEC')
-        layout.menu('MOL_MT_ADD_NODE_MENU_BONDS', text='Bonds', icon = 'FIXED_SIZE')
-        layout.menu('MOL_MT_ADD_NODE_MENU_SELECTIONS', text='Selection', icon_value=256)
-        layout.menu('MOL_MT_ADD_NODE_MENU_ANIMATION', text='Animation', icon_value=409)
-        layout.menu('MOL_MT_ADD_NODE_MENU_ASSEMBLY', text='Assemblies', icon = 'GROUP_VERTEX')
-        # layout.menu('MOL_MT_ADD_NODE_MENU_MEMBRANES', text='Membranes', icon_value=248)
-        layout.menu('MOL_MT_ADD_NODE_MENU_DNA', text='DNA', icon='GP_SELECT_BETWEEN_STROKES')
-        layout.menu('MOL_MT_ADD_NODE_MENU_UTILITIES', text='Utilities', icon_value=92)
+        layout.menu('MOL_MT_ADD_NODE_MENU_SYLING', 
+                    text='Style', icon_value=77)
+        layout.menu('MOL_MT_ADD_NODE_MENU_COLOR', 
+                    text='Color', icon = 'COLORSET_07_VEC')
+        layout.menu('MOL_MT_ADD_NODE_MENU_BONDS', 
+                    text='Bonds', icon = 'FIXED_SIZE')
+        layout.menu('MOL_MT_ADD_NODE_MENU_SELECTIONS', 
+                    text='Selection', icon_value=256)
+        layout.menu('MOL_MT_ADD_NODE_MENU_ANIMATION', 
+                    text='Animation', icon_value=409)
+        layout.menu('MOL_MT_ADD_NODE_MENU_ASSEMBLY', 
+                    text='Assemblies', icon = 'GROUP_VERTEX')
+        layout.menu('MOL_MT_ADD_NODE_MENU_DNA', 
+                    text='DNA', icon='GP_SELECT_BETWEEN_STROKES')
+        layout.menu('MOL_MT_ADD_NODE_MENU_UTILITIES', 
+                    text='Utilities', icon_value=92)
 
 def mol_add_node_menu(self, context):
     if ('GeometryNodeTree' == bpy.context.area.spaces[0].tree_type):